feat(azure): delegate-domain support — create DNS zone, refactor cert flow
Implements PrepareDomainDelegation for the Azure provider so each project
gets a dedicated public DNS zone in Azure DNS, returning the zone's name
servers for Fabric to delegate. Mirrors the GCP provider's approach.

Headline pieces:

- pkg/clouds/azure/dns: armdns wrapper with EnsureZoneExists (idempotent
  get-then-create), matching GCP's EnsureDNSZoneExists.

- pkg/cli/client/byoc/azure/byoc.go: PrepareDomainDelegation creates the
  zone in the per-project-stack resource group (ensuring the RG exists
  via the existing idempotent CreateResourceGroup), returns the NS records,
  and stores the zone name for parity with the GCP provider's
  delegateDomainZone field. HasDelegatedSubdomain returns true so destroy
  calls fabric.DeleteSubdomainZone. DOMAIN env var is forwarded to the CD
  task via the cdCommand struct, mirroring the AWS provider's pattern.

- pkg/clouds/azure/aca/cert.go: moves the existing IssueCert flow from
  pkg/cli/client/byoc/azure/cert.go into the cloud-SDK layer so both the
  CLI's `defang cert generate` and the CD task (separate repo) can call
  it directly without crossing the byoc → cloud-SDK boundary. The
  *ByocAzure.IssueCert method is now a thin wrapper. Pure refactor for
  the CLI's BYOD path; idempotency of every ARM step is preserved.

- pkg/cli/client/byoc/common.go: when DEFANG_PULUMI_DIR runs CD locally,
  forward AZURE_STORAGE_{KEY,ACCOUNT,SAS_TOKEN} and AZURE_TENANT_ID from
  the host shell into the subprocess so Pulumi's azblob backend can
  authenticate directly. Avoids DefaultAzureCredential falling through
  to AzureCLICredential, which times out under nix's slow `az` startup.
  No-op for the ACA-job path, which uses managed identity in-cloud.

The companion pulumi-defang change adds the per-service CNAME + asuid TXT
records (Pulumi-managed) and the CD-side hook that calls aca.IssueCert to
register the customDomain + issue ManagedCertificate + bind SniEnabled,
completing the delegate-domain flow.

Verified end-to-end on Azure westus3 with the nextjs sample: fresh deploy
provisions zone → records → cert + binding in ~9 min; HTTPS serves a
valid DigiCert-issued cert for app.<delegate-domain> in ~0.6s response.
