+++ src/pkg/cli/client/byoc/azure/byoc.go
@@ -23,6 +23,7 @@ import (
 	"github.com/DefangLabs/defang/src/pkg/clouds/azure/aca"
 	"github.com/DefangLabs/defang/src/pkg/clouds/azure/acr"
 	"github.com/DefangLabs/defang/src/pkg/clouds/azure/cd"
+	azuredns "github.com/DefangLabs/defang/src/pkg/clouds/azure/dns"
 	"github.com/DefangLabs/defang/src/pkg/clouds/azure/keyvault"
 	defanghttp "github.com/DefangLabs/defang/src/pkg/http"
 	"github.com/DefangLabs/defang/src/pkg/term"
@@ -44,6 +45,10 @@ type ByocAzure struct {
 	cdRunID string
 	cdEtag  string
 	cdStart time.Time
+
+	// delegateDomainZone records the delegate domain whose public DNS zone
+	// PrepareDomainDelegation created, mirroring the GCP provider.
+	delegateDomainZone string
 }
 
 var _ client.Provider = (*ByocAzure)(nil)
@@ -435,12 +440,13 @@ func (b *ByocAzure) Deploy(ctx context.Context, req *client.DeployRequest) (*cli
 }
 
 type cdCommand struct {
-	command   []string
-	etag      types.ETag
-	mode      defangv1.DeploymentMode
-	project   string
-	statesUrl string
-	eventsUrl string
+	command        []string
+	etag           types.ETag
+	mode           defangv1.DeploymentMode
+	project        string
+	statesUrl      string
+	eventsUrl      string
+	delegateDomain string // forwarded to CD as DOMAIN; empty when deploying without delegation
 }
 
 func (b *ByocAzure) runCdCommand(ctx context.Context, cmd cdCommand) (string, error) {
@@ -461,6 +467,14 @@ func (b *ByocAzure) runCdCommand(ctx context.Context, cmd cdCommand) (string, er
 
 	if cmd.eventsUrl != "" {
 		env["DEFANG_EVENTS_UPLOAD_URL"] = cmd.eventsUrl
+	}
+
+	// DOMAIN is the project delegate domain (e.g. "myproj-stack.tenant.defang.app");
+	// the CD program reads it and sets defang:domain Pulumi config, which the
+	// Azure inline program forwards to ProjectInputs.Domain so per-service
+	// DNS records and managed certs land in the right zone. Mirrors AWS.
+	if cmd.delegateDomain != "" {
+		env["DOMAIN"] = cmd.delegateDomain
 	}
 
 	if os.Getenv("DEFANG_PULUMI_DIR") != "" {
@@ -546,12 +560,13 @@ func (b *ByocAzure) deploy(ctx context.Context, req *client.DeployRequest, verb 
 
 	cdStart := time.Now()
 	execName, err := b.runCdCommand(ctx, cdCommand{
-		command:   []string{verb, payload},
-		etag:      etag,
-		mode:      req.Mode,
-		project:   project.Name,
-		statesUrl: req.StatesUrl,
-		eventsUrl: req.EventsUrl,
+		command:        []string{verb, payload},
+		etag:           etag,
+		mode:           req.Mode,
+		project:        project.Name,
+		statesUrl:      req.StatesUrl,
+		eventsUrl:      req.EventsUrl,
+		delegateDomain: req.DelegateDomain,
 	})
 	if err != nil {
 		return nil, err
@@ -708,17 +723,48 @@ func (b *ByocAzure) ListConfig(ctx context.Context, req *defangv1.ListConfigsReq
 	return &defangv1.Secrets{Names: names}, nil
 }
 
-// PrepareDomainDelegation implements client.Provider.
-func (b *ByocAzure) PrepareDomainDelegation(context.Context, client.PrepareDomainDelegationRequest) (*client.PrepareDomainDelegationResponse, error) {
-	return nil, nil // TODO: implement domain delegation for Azure
+// PrepareDomainDelegation implements client.Provider. It creates a public
+// Azure DNS zone for the delegate domain and returns the zone's authoritative
+// name servers so Fabric can point NS records at it, mirroring the GCP
+// provider. The zone lives in the per-project-stack resource group so it is
+// owned by this deployment and torn down with it. Azure has no equivalent of
+// Route53's reusable delegation sets, so DelegationSetId is left empty (as
+// with GCP).
+func (b *ByocAzure) PrepareDomainDelegation(ctx context.Context, req client.PrepareDomainDelegationRequest) (*client.PrepareDomainDelegationResponse, error) {
+	defer term.Timing()()
+	if req.DelegateDomain == "" {
+		return nil, nil // no delegate domain configured; nothing to delegate
+	}
+	if err := b.setUpLocation(); err != nil {
+		return nil, err
+	}
+	term.Debugf("Preparing domain delegation for %s", req.DelegateDomain)
+
+	// The zone needs a home before it can be created. On a fresh deploy Pulumi
+	// has not created the project resource group yet, so ensure it exists
+	// (idempotent CreateOrUpdate) the same way config setup does.
+	rgName := b.projectResourceGroupName(req.Project)
+	if err := b.driver.CreateResourceGroup(ctx, rgName); err != nil {
+		return nil, err
+	}
+
+	nameServers, err := azuredns.New(rgName, b.driver.Azure).EnsureZoneExists(ctx, req.DelegateDomain)
+	if err != nil {
+		return nil, err
+	}
+	b.delegateDomainZone = req.DelegateDomain
+	term.Debugf("Zone %s ready with nameservers %v", req.DelegateDomain, nameServers)
+	return &client.PrepareDomainDelegationResponse{
+		NameServers: nameServers,
+	}, nil
 }
 
-// HasDelegatedSubdomain implements client.Provider. Azure does not yet
-// implement domain delegation, so there is no subdomain zone to delete on
-// destroy. Flip to true (or remove the override) once PrepareDomainDelegation
-// is wired up.
+// HasDelegatedSubdomain implements client.Provider. Azure delegates a
+// subdomain zone during deploy (PrepareDomainDelegation +
+// CreateDelegateSubdomainZone), so the matching DeleteSubdomainZone must run
+// on destroy.
 func (*ByocAzure) HasDelegatedSubdomain() bool {
-	return false
+	return true
 }
 
 // Preview implements client.Provider.
