+++ src/pkg/clouds/azure/aca/cert_test.go
@@ -1,4 +1,4 @@
-package azure
+package aca
 
 import (
 	"errors"
